Add files via upload
this is my first commit on the javaPrograms, in this problem we have to add the elements in the list and print the result, make sure you read the instructions carefully while solving a problem on hackerrank etc because i didn't and was cofused that why is my output coming wrong. but later realized it and corrected myself.