Add files via upload
in this problem we have to find the absolute difference between both the diagonals of a Square ArrayList Matrix, this one is a little bit tricky bcuz arraylist matrix is not the same as array matrix. I just used one loop to iterate the index of the first diagonal and the second diagonal. Its important to note that the output requires an absolute value(ie: positive or negative should not affect it) so we use Math.abs() function of the math utility so that we get an absolute value of the calculation